Rename internal strings accessor module to match runtime
