+++ pandas-stubs/core/strings/accessor.pyi
